Refactor of styles components
