+++ src/styles/base.scss
@@ -1,1 +1,14 @@
 $main-color: rgb(28 13 93);
+
+
+.commonButton {
+    border-radius: 6px;
+    text-align: center;
+    font-weight: bold;
+    border: solid 2px $main-color;
+}
+
+.container_column {
+    display: flex;
+    flex-direction: column;
+}
