+++ src/styles/components/registration/InformativeMessage.module.scss
@@ -1,4 +1,4 @@
-@use '../base';
+@use '../../base';
 
 
 .title{
